fix: 운영 시간 useStoreTimeSetUp 훅으로 관리
